+++ src/registry/mod.rs
@@ -32,7 +32,7 @@ pub mod prometheus_0_14;
 
 #[cfg(feature = "prometheus-client")]
 #[cfg_attr(docsrs, doc(cfg(feature = "prometheus-client")))]
-pub use prometheus_client_0_23 as prometheus_client;
+pub use prometheus_client_0_24 as prometheus_client;
 
 /// Prometheus metrics components.
 #[cfg(feature = "prometheus-client_0_22")]
@@ -43,6 +43,11 @@ pub mod prometheus_client_0_22;
 #[cfg(feature = "prometheus-client_0_23")]
 #[cfg_attr(docsrs, doc(cfg(feature = "prometheus-client_0_23")))]
 pub mod prometheus_client_0_23;
+
+/// Prometheus metrics components.
+#[cfg(feature = "prometheus-client_0_24")]
+#[cfg_attr(docsrs, doc(cfg(feature = "prometheus-client_0_24")))]
+pub mod prometheus_client_0_24;
 
 #[cfg(feature = "opentelemetry")]
 #[cfg_attr(docsrs, doc(cfg(feature = "opentelemetry")))]