+++ CHANGELOG.md
@@ -15,6 +15,6 @@ Supports the following metrics backends:
 
 - Prometheus
     - [`prometheus`](https://crates.io/crates/prometheus): `0.13`, `0.14`
-    - [`prometheus-client`](https://crates.io/crates/prometheus-client): `0.22`, `0.23`
+    - [`prometheus-client`](https://crates.io/crates/prometheus-client): `0.22`, `0.23`, `0.24`
 - OpenTemeletry Metrics
     - [`opentelemetry`](https://crates.io/crates/opentelemetry): `0.26`, `0.27`, `0.28`, `0.29`